requirements.txt
# Core ML/NLP libraries
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# ML & Deep Learning
torch>=2.0.0
transformers>=4.30.0
bitsandbytes>=0.41.0
accelerate>=0.20.0

# Semantic Search & Retrieval
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# ML Models for Evaluation
scikit-learn>=1.3.0
xgboost>=2.0.0

# Feature Extraction
textblob>=0.17.0
vaderSentiment>=3.3.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.12.0
scikit-learn>=1.3.0

# NLP utilities
nltk>=3.8.0
textblob>=0.17.0

# Model inference optimization
bitsandbytes>=0.41.0
optimum>=1.14.0

# UI
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.1.0

# API & deployment
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
loguru>=0.7.2

# Data validation
jsonschema>=4.19.0
